@@ -1,6 +1,6 @@
 {
   "name": "ngx-crud-meta-form",
-  "version": "0.0.21",
+  "version": "0.0.22",
   "description": "Library to support the meta info controlled edit of base data",
   "author": "lunorsys GmbH, www.lunorsys.de",
   "repository": {
@@ -21,39 +21,39 @@
   ],
   "license": "lunorsys license",
   "peerDependencies": {
-    "@angular-devkit/schematics": "^9.1.7",
-    "@angular/animations": "^9.1.9",
-    "@angular/cdk": "^9.2.3",
-    "@angular/cli": "^9.1.7",
-    "@angular/common": "^9.1.9",
-    "@angular/compiler": "^9.1.9",
-    "@angular/core": "^9.1.9",
-    "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "^9.1.9",
-    "@angular/material": "^9.2.3",
-    "@angular/material-moment-adapter": "^9.2.3",
-    "@angular/platform-browser": "^9.1.9",
-    "@angular/platform-browser-dynamic": "^9.1.9",
-    "@angular/router": "^9.1.9",
-    "@code-art/angular-globalize": "^9.0.1",
-    "@types/node": "^14.0.10",
+    "@angular-devkit/schematics": "^10.0.3",
+    "@angular/animations": "^10.0.4",
+    "@angular/cdk": "^10.1.0",
+    "@angular/cli": "^10.0.3",
+    "@angular/common": "^10.0.4",
+    "@angular/compiler": "^10.0.4",
+    "@angular/core": "^10.0.4",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "^10.0.4",
+    "@angular/material": "^10.1.0",
+    "@angular/material-moment-adapter": "^10.1.0",
+    "@angular/platform-browser": "^10.0.4",
+    "@angular/platform-browser-dynamic": "^10.0.4",
+    "@angular/router": "^10.0.4",
+    "@code-art/angular-globalize": "^9.0.4",
+    "@types/node": "^14.0.24",
     "cldr": "^5.6.0",
     "cldr-data": "^36.0.0",
-    "codelyzer": "^5.2.2",
+    "codelyzer": "^6.0.0",
     "core-js": "^3.6.5",
-    "eslint": "^7.1.0",
     "fstream": "^1.0.12",
     "globalize": "^1.5.0",
-    "lodash": "^4.17.15",
-    "moment": "^2.26.0",
+    "lodash": "^4.17.19",
+    "moment": "^2.27.0",
     "node-sass": "^4.14.1",
-    "rxjs": "^6.5.5",
-    "rxjs-compat": "^6.5.5",
+    "rxjs": "^6.6.0",
+    "rxjs-compat": "^6.6.0",
     "rxjs-tslint": "^0.1.8",
     "sockjs-client": "^1.4.0",
-    "tslib": "^1.10.0",
     "tslint": "^6.1.2",
     "zone.js": "^0.10.3"
   },
-  "dependencies": {}
+  "dependencies": {
+    "tslib": "^2.0.0"
+  }
 }
